tweak sublist
Stolen from
https://exercism.org/tracks/clojure/exercises/sublist/solutions/amirci,
but with comments.

Note that it's OK for `sublist?` to return `true` or `nil`, because we
wrap it with a `cond` in `classify`. That is, if `sublist?` evaluates to
`nil` (instead of false) it's OK because the condition will be skipped
to the next one in `cond` because `cond` only cares about truthy
branches. Coming from Haskell where boolean functions must return true
or false, this feels a bit dirty but it gets the job done.