tweak raindrops
For `cond->`, there is no short-circuit behavior like `cond`. So using
the same `:else` keyword (truthy value) at the end is a little bit
misleading because it suggests that it will only get executed if none of
the previous conditionals are true. But `cond->` evaluates *all* truthy
conditionals, so using an `:always` keyword is more explicit.